Centralize objectNode logic in one file
