+++ packages/dds/tree/src/simple-tree/arrayNode.ts
