+++ packages/dds/tree/src/simple-tree/index.ts
@@ -24,12 +24,9 @@ export {
 	ApplyKind,
 	FieldKind,
 	FieldSchema,
-	InsertableObjectFromSchemaRecord,
 	InsertableTreeFieldFromImplicitField,
 	InsertableTypedNode,
 	NodeBuilderData,
-	ObjectFromSchemaRecord,
-	TreeObjectNode,
 	type FieldProps,
 } from "./schemaTypes.js";
 export { SchemaFactory, type ScopedSchemaName } from "./schemaFactory.js";
@@ -70,4 +67,9 @@ export {
 } from "./testRecursiveDomain.js";
 
 export { TreeNode, Unhydrated } from "./types.js";
-export { TreeArrayNode, IterableTreeArrayContent, TreeArrayNodeBase } from "./treeArrayNode.js";
+export { TreeArrayNode, IterableTreeArrayContent, TreeArrayNodeBase } from "./arrayNode.js";
+export {
+	InsertableObjectFromSchemaRecord,
+	ObjectFromSchemaRecord,
+	TreeObjectNode,
+} from "./objectNode.js";
