+++ packages/dds/tree/src/simple-tree/schemaFactoryRecursive.ts
@@ -5,6 +5,7 @@
 
 import { FlexTreeNode, Unenforced, isFlexTreeNode } from "../feature-libraries/index.js";
 import { RestrictiveReadonlyRecord } from "../util/index.js";
+import { InsertableObjectFromSchemaRecord } from "./objectNode.js";
 
 import { SchemaFactory, type ScopedSchemaName } from "./schemaFactory.js";
 import {
@@ -12,7 +13,6 @@ import {
 	FieldSchema,
 	ImplicitAllowedTypes,
 	ImplicitFieldSchema,
-	InsertableObjectFromSchemaRecord,
 	InsertableTreeNodeFromImplicitAllowedTypes,
 	NodeKind,
 	TreeNodeSchema,
